convert configuration to dict & refactor implementation details into separate header
