+++ tests/test_config.py
@@ -542,7 +542,7 @@ def test_group():
     assert cfg['lvl1.lvl2.flt'] == pytest.approx(2)
 
 
-def test_dict():
+def test_dict_set():
     cfg_toml = pyc.load_toml_str("""
         int = 1
         
@@ -576,6 +576,9 @@ def test_dict():
     assert isinstance(cfg['dict'], type(cfg))
     cfg = cfg['dict']
     assert cfg == cfg_toml
+
+    tmp = cfg.to_dict()
+    assert tmp == pydict
 
     assert 2 == len(cfg)
     assert 1 == cfg['int']
@@ -611,6 +614,71 @@ def test_dict():
     lst = [1, 2, {"foo.invalid": "bar", "3": "value"}]
     with pytest.raises(pyc.TypeError):
         cfg['mixed-lst'] = lst
+
+
+def test_dict_get():
+    cfg = pyc.load_toml_str("""
+        [scalars]
+        int = 1
+        flt = 1e-17
+        flag = true
+        str = 'value'
+
+        [lists]
+        nums = [1, 2, 3.123]
+        mixed = [true, 42, 'value']
+        nested = [
+            [1, 2, 3],
+            17,
+            { str = 'value' }
+        ]
+
+        [tables]
+        t1 = { name = 'value' }
+        t2 = { name = 'foo', t2-1 = { name = 'bar' }}
+
+        [dates]
+        day = 2023-03-19
+        time = 19:48:00
+        dt = 2023-03-19T19:48:00+01:00
+        """)
+    d = cfg.to_dict()
+    assert len(d) == 4
+    assert 'scalars' in d
+    assert isinstance(d['scalars'], dict)
+    assert isinstance(d['scalars']['int'], int)
+    assert isinstance(d['scalars']['flt'], float)
+    assert isinstance(d['scalars']['flag'], bool)
+    assert isinstance(d['scalars']['str'], str)
+
+    assert 'lists' in d
+    assert isinstance(d['lists'], dict)
+    assert isinstance(d['lists']['nums'], list)
+    assert [1, 2, 3.123] == d['lists']['nums']
+
+    assert isinstance(d['lists']['mixed'], list)
+    assert [True, 42, 'value'] == d['lists']['mixed']
+
+    assert isinstance(d['lists']['nested'], list)
+    assert isinstance(d['lists']['nested'][0], list)
+    assert [1, 2, 3] == d['lists']['nested'][0]
+    assert isinstance(d['lists']['nested'][1], int)
+    assert isinstance(d['lists']['nested'][2], dict)
+
+    assert 'tables' in d
+    assert isinstance(d['tables'], dict)
+    assert isinstance(d['tables']['t1'], dict)
+    assert isinstance(d['tables']['t1']['name'], str)
+    assert isinstance(d['tables']['t2'], dict)
+    assert isinstance(d['tables']['t2']['name'], str)
+    assert isinstance(d['tables']['t2']['t2-1'], dict)
+    assert isinstance(d['tables']['t2']['t2-1']['name'], str)
+
+    assert 'dates' in d
+    assert isinstance(d['dates'], dict)
+    assert isinstance(d['dates']['day'], datetime.date)
+    assert isinstance(d['dates']['time'], datetime.time)
+    assert isinstance(d['dates']['dt'], datetime.datetime)
 
 
 def test_size():
@@ -725,6 +793,7 @@ def test_keys():
     cfg['valid'] = 3
     assert 'valid' in cfg
     assert 3 == cfg['valid']
+
 
 def test_none():
     cfg = pyc.load_toml_str("""
